requirements.txt
discord.py[voice]
psutil
jishaku
feedparser
aiomysql
youtube_dl
ffmpeg
googleapiclient
